+++ application/src/components/locales/en/Call-for-papers.vue
@@ -39,8 +39,8 @@ export default {
     ]"
     :milestones="[
       { label: 'Extended submission deadline', date: '2026-06-21' },
-      { label: 'Notification of acceptance', date: '2026-08-07' },
-      { label: 'Camera-ready final submissions', date: '2026-09-01' },
+      { label: 'Notification of acceptance', date: '2026-08-22' },
+      { label: 'Camera-ready final submissions', date: '2026-09-16' },
       {
         label: 'MexIHC 2026 starts',
         date: '2026-10-27',
@@ -120,8 +120,8 @@ export default {
   <h2 id="cfp-dates">Important Dates</h2>
   <ul>
     <li><del>June 1, 2026</del> <del>June 15, 2026</del> <strong>June 21, 2026</strong> – Deadline for submissions <strong>(extended)</strong></li>
-    <li>Notification of acceptance: August 7, 2026</li>
-    <li>Camera-ready final submissions: September 1, 2026</li>
+    <li>Notification of acceptance: August 22, 2026</li>
+    <li>Camera-ready final submissions: September 16, 2026</li>
     <li>MexIHC 2026: October 27-30, 2026</li>
   </ul>